chore: check whether react-query wrapper has accessToken present (#882)
Because

- Better debug

This commit

- check whether react-query wrapper has accessToken present

+++ packages/toolkit/src/lib/react-query-service/model/useDeleteUserModel.ts
@@ -13,7 +13,12 @@ export const useDeleteModel = () => {
       modelName: string;
       accessToken: Nullable<string>;
     }) => {
+      if (!accessToken) {
+        return Promise.reject(new Error("accessToken not provided"));
+      }
+
       await deleteUserModelMutation({ modelName, accessToken });
+
       return Promise.resolve({ modelName, accessToken });
     },
     {
